Hook up automation to quit all neovim instances prior to exiting
